Tested with the new queue class.
The main rendering still works without a hitch, which is great.
The close mechanic is vastly superior to the previous version of
this code, which had to setup a sentinel element per worker thread.
An alternate design that I would like to try which would exercise
the queue more would be to leave the queue and workers and let it
all happen live so that the "circular" part of the queueing is
actually exercised.